backend/requirements.txt
accelerate==0.30.1
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
langchain-community==0.2.1
multiwordnet==0.1.5
requests==2.31.0
torch==2.2.2
transformers==4.40.1
Werkzeug==3.0.2